backend/requirements.txt
Flask
Flask-CORS
Flask-MySQLdb
mysql-connector-python
python-dotenv
werkzeug